Migrate from Swarms to LangGraph with full Gemini 2.5 Flash support
Major Changes:
- Replace Swarms framework with LangGraph for better state management
- Implement proper message formatting for Gemini API compatibility
- Add comprehensive agent system with 8 specialized medical AI agents
- Create state-managed workflow with TypedDict and proper tool integration
- Support for multiple LLM providers (OpenAI, Anthropic, Google Gemini)
- Add budget tracking and cost management features
- Implement diagnostic workflow with proper agent handoffs
- Add comprehensive error handling and rate limit management

Technical Implementation:
- New state management system (state.py) with DiagnosticState
- Modular agent architecture (agents/ directory)
- Structured tool integration (tools.py) with proper Command returns
- Graph-based workflow orchestration (graph.py)
- Message format compatibility fixes for Gemini API
- Updated dependencies and configuration

Testing:
- Verified with Gemini 2.5 Flash, 2.0 Flash Exp, and 1.5 Flash
- Complete diagnostic workflow execution
- Multi-agent coordination and tool calling
- Proper state persistence and checkpointing
