+++ .env.example
@@ -1,4 +1,6 @@
 WORKSPACE_DIR=""
 OPENAI_API_KEY=""
 ANTHROPIC_API_KEY=""
+GOOGLE_API_KEY=""
 GEMINI_API_KEY=""
+
